+++ api/weatherCompare/weatherCompare.go
@@ -14,56 +14,13 @@ import (
 	"time"
 )
 
-// data structure stores weather data for a location.
-type data struct {
-	Longitude float64 `json:"longitude"`
-	Latitude  float64 `json:"latitude"`
-	Location  string  `json:"location"`
-	Updated   string  `json:"updated"`
-	Instant   struct {
-		AirTemperature      float64 `json:"air_temperature"`
-		CloudAreaFraction   float64 `json:"cloud_area_fraction"`
-		DewPointTemperature float64 `json:"dew_point_temperature"`
-		RelativeHumidity    float64 `json:"relative_humidity"`
-		WindSpeed           float64 `json:"wind_speed"`
-		WindSpeedOfGust     float64 `json:"wind_speed_of_gust"`
-		PrecipitationAmount float64 `json:"precipitation_amount"`
-	} `json:"instant"`
-	Predicted struct {
-		AirTemperatureMax          float64 `json:"air_temperature_max"`
-		AirTemperatureMin          float64 `json:"air_temperature_min"`
-		PrecipitationAmount        float64 `json:"precipitation_amount"`
-		PrecipitationAmountMax     float64 `json:"precipitation_amount_max"`
-		PrecipitationAmountMin     float64 `json:"precipitation_amount_min"`
-		ProbabilityOfPrecipitation float64 `json:"probability_of_precipitation"`
-	} `json:"predicted"`
-}
-
-// locationInfo structure stores all comparison locations information.
-type locationInfo struct {
-	Location  string
-	Longitude float64
-	Latitude  float64
-}
-
-// WeatherCompare structure stores current and predicted weather data comparisons for different locations.
-//
-// Functionality: Handler, get
-type WeatherCompare struct {
-	Longitude  float64 			 `json:"longitude"`
-	Latitude   float64 			 `json:"latitude"`
-	Location   string  			 `json:"location"`
-	Updated    string  			 `json:"updated"`
-	Timeseries map[string][]data `json:"timeseries"`
-}
-
 // Handler will handle http request for REST service.
 func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Request) {
 	//parse url and branch if an error occurred
 	arrPath := strings.Split(r.URL.Path, "/")
 	if len(arrPath) != 7 {
 		debug.ErrorMessage.Update(
-			http.StatusBadRequest, 
+			http.StatusBadRequest,
 			"WeatherCompare.Handler() -> Parsing URL",
 			"url validation: either too many or too few arguments in url path",
 			"URL format. Expected format: '.../place'. Example: '.../oslo'",
@@ -71,28 +28,16 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 		debug.ErrorMessage.Print(w)
 		return
 	}
+	//get main location information and branch if an error occurred
 	mainLocation := arrPath[5]
-	compareLocations := arrPath[6]
-	//get locations to compare and branch if there are none
-	arrCompareLocations := strings.Split(compareLocations, ";")
-	if len(arrCompareLocations) < 1 {
-		debug.ErrorMessage.Update(
-			http.StatusBadRequest, 
-			"WeatherCompare.Handler() -> Getting locations to compare",
-			"url validation: not enough locations to compare to",
-			"URL format. Expected format: '.../main_place/place1;place2;...'. Example: '.../bergen/oslo;stavanger'",
-		)
-		debug.ErrorMessage.Print(w)
-		return
-	}
 	var mainLocationCoords geoCoords.LocationCoords
 	status, err := mainLocationCoords.Handler(mainLocation)
 	if err != nil {
 		debug.ErrorMessage.Update(
-			status, 
+			status,
 			"WeatherCompare.Handler() -> LocationCoords.Handler() -> Getting main location info",
 			err.Error(),
-			"UnkInstantn",
+			"Unknown",
 		)
 		debug.ErrorMessage.Print(w)
 		return
@@ -100,16 +45,30 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 	weatherCompare.Longitude = mainLocationCoords.Longitude
 	weatherCompare.Latitude = mainLocationCoords.Latitude
 	weatherCompare.Location = mainLocationCoords.Address
+	//get locations to compare and branch if there are none
+	compareLocations := arrPath[6]
+	arrCompareLocations := strings.Split(compareLocations, ";")
+	if len(arrCompareLocations) < 1 {
+		debug.ErrorMessage.Update(
+			http.StatusBadRequest,
+			"WeatherCompare.Handler() -> Getting locations to compare",
+			"url validation: not enough locations to compare to",
+			"URL format. Expected format: '.../main_place/place1;place2;...'. Example: '.../bergen/oslo;stavanger'",
+		)
+		debug.ErrorMessage.Print(w)
+		return
+	}
+	//get compare locations information and branch if an error occurred
 	var arrCoordinates []locationInfo
 	for _, location := range arrCompareLocations {
 		var locationCoords geoCoords.LocationCoords
 		status, err := locationCoords.Handler(location)
 		if err != nil {
 			debug.ErrorMessage.Update(
-				status, 
+				status,
 				"WeatherCompare.Handler() -> LocationCoords.Handler() -> Getting comparison location info",
 				err.Error(),
-				"UnkInstantn",
+				"Unknown",
 			)
 			debug.ErrorMessage.Print(w)
 			return
@@ -124,7 +83,7 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 	arrParam, err := url.ParseQuery(r.URL.RawQuery)
 	if err != nil {
 		debug.ErrorMessage.Update(
-			http.StatusInternalServerError, 
+			http.StatusInternalServerError,
 			"WeatherCompare.Handler() -> Validating URL parameters",
 			err.Error(),
 			"Unknown",
@@ -142,7 +101,7 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 			_, err = time.Parse("2006-01-02", date)
 			if err != nil {
 				debug.ErrorMessage.Update(
-					http.StatusBadRequest, 
+					http.StatusBadRequest,
 					"WeatherCompare.Handler() -> Validating URL parameters",
 					err.Error(),
 					"Date doesn't match YYYY-MM-DD format. Example: 2021-04-26",
@@ -152,7 +111,7 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 			}
 		} else {
 			debug.ErrorMessage.Update(
-				http.StatusBadRequest, 
+				http.StatusBadRequest,
 				"WeatherCompare.Handler() -> Validating URL parameters",
 				"url validation: unknown parameter",
 				"Unknown",
@@ -162,10 +121,10 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 		}
 	}
 	//get weather data and branch if an error occurred
-	status, err = weatherCompare.get(mainLocationCoords.Latitude, mainLocationCoords.Longitude, arrCoordinates, date)
+	status, err = weatherCompare.get(weatherCompare.Latitude, weatherCompare.Longitude, arrCoordinates, date)
 	if err != nil {
 		debug.ErrorMessage.Update(
-			status, 
+			status,
 			"WeatherCompare.Handler() -> WeatherCompare.get() -> Getting data",
 			err.Error(),
 			"UnkInstantn",
@@ -173,14 +132,12 @@ func (weatherCompare *WeatherCompare) Handler(w http.ResponseWriter, r *http.Req
 		debug.ErrorMessage.Print(w)
 		return
 	}
-	//update header to JSON and set HTTP code
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
 	//send output to user and branch if an error occured
+	w.Header().Set("Content-Type", "application/json")
 	err = json.NewEncoder(w).Encode(weatherCompare)
 	if err != nil {
 		debug.ErrorMessage.Update(
-			http.StatusInternalServerError, 
+			http.StatusInternalServerError,
 			"WeatherCompare.Handler() -> Sending data to user",
 			err.Error(),
 			"UnkInstantn",
@@ -194,15 +151,13 @@ func (weatherCompare *WeatherCompare) get(lat float64, lon float64, arrCoordinat
 	//convert coordinates to string
 	strLat := fmt.Sprintf("%f", lat)
 	strLon := fmt.Sprintf("%f", lon)
-	//get weather data and branch if an error occurred
+	//get weather data for main location and branch if an error occurred
 	var mainWeatherData weatherData.WeatherData
 	status, err := mainWeatherData.Handler(strLat, strLon)
 	if err != nil {
 		return status, err
 	}
-	//set data in structure
-	weatherCompare.Updated = mainWeatherData.Updated
-	//validate date
+	//check if there is weather data for inputted date
 	if _, ok := mainWeatherData.Timeseries[date]; !ok {
 		return http.StatusBadRequest, errors.New("invalid date: can't find weather data for inputted date")
 	}
@@ -225,7 +180,6 @@ func (weatherCompare *WeatherCompare) get(lat float64, lon float64, arrCoordinat
 		data.Latitude = coordinates.Latitude
 		data.Location = coordinates.Location
 		data.Updated = weatherData.Updated
-
 		data.Instant.AirTemperature = fun.LimitDecimals(weatherData.Timeseries[date].Instant.AirTemperature - mainWeatherData.Timeseries[date].Instant.AirTemperature)
 		data.Instant.CloudAreaFraction = fun.LimitDecimals(weatherData.Timeseries[date].Instant.CloudAreaFraction - mainWeatherData.Timeseries[date].Instant.CloudAreaFraction)
 		data.Instant.DewPointTemperature = fun.LimitDecimals(weatherData.Timeseries[date].Instant.DewPointTemperature - mainWeatherData.Timeseries[date].Instant.DewPointTemperature)
@@ -233,16 +187,15 @@ func (weatherCompare *WeatherCompare) get(lat float64, lon float64, arrCoordinat
 		data.Instant.WindSpeed = fun.LimitDecimals(weatherData.Timeseries[date].Instant.WindSpeed - mainWeatherData.Timeseries[date].Instant.WindSpeed)
 		data.Instant.WindSpeedOfGust = fun.LimitDecimals(weatherData.Timeseries[date].Instant.WindSpeedOfGust - mainWeatherData.Timeseries[date].Instant.WindSpeedOfGust)
 		data.Instant.PrecipitationAmount = fun.LimitDecimals(weatherData.Timeseries[date].Instant.PrecipitationAmount - mainWeatherData.Timeseries[date].Instant.PrecipitationAmount)
-
 		data.Predicted.AirTemperatureMax = fun.LimitDecimals(weatherData.Timeseries[date].Predicted.AirTemperatureMax - mainWeatherData.Timeseries[date].Predicted.AirTemperatureMax)
 		data.Predicted.AirTemperatureMin = fun.LimitDecimals(weatherData.Timeseries[date].Predicted.AirTemperatureMin - mainWeatherData.Timeseries[date].Predicted.AirTemperatureMin)
 		data.Predicted.PrecipitationAmount = fun.LimitDecimals(weatherData.Timeseries[date].Predicted.PrecipitationAmount - mainWeatherData.Timeseries[date].Predicted.PrecipitationAmount)
 		data.Predicted.PrecipitationAmountMax = fun.LimitDecimals(weatherData.Timeseries[date].Predicted.PrecipitationAmountMax - mainWeatherData.Timeseries[date].Predicted.PrecipitationAmountMax)
 		data.Predicted.PrecipitationAmountMin = fun.LimitDecimals(weatherData.Timeseries[date].Predicted.PrecipitationAmountMin - mainWeatherData.Timeseries[date].Predicted.PrecipitationAmountMin)
 		data.Predicted.ProbabilityOfPrecipitation = fun.LimitDecimals(weatherData.Timeseries[date].Predicted.ProbabilityOfPrecipitation - mainWeatherData.Timeseries[date].Predicted.ProbabilityOfPrecipitation)
-		//append data to array
 		dataRange = append(dataRange, data)
 	}
 	weatherCompare.Timeseries[date] = dataRange
+	weatherCompare.Updated = mainWeatherData.Updated
 	return http.StatusOK, nil
 }