+++ api/weatherData/weatherData.go
@@ -1,70 +1,41 @@
 package weatherData
 
 import (
+	"errors"
 	"main/db"
 	"net/http"
 	"time"
 )
 
-// Timeseries stores current and predicted weather data for a day
-type Timeseries struct {
-	Instant struct {
-		AirTemperature      float64 `json:"air_temperature"`
-		CloudAreaFraction   float64 `json:"cloud_area_fraction"`
-		DewPointTemperature	float64 `json:"dew_point_temperature"`
-		RelativeHumidity    float64 `json:"relative_humidity"`
-		WindFromDirection   float64 `json:"wind_from_direction"`
-		WindSpeed           float64 `json:"wind_speed"`
-		WindSpeedOfGust     float64 `json:"wind_speed_of_gust"`
-		PrecipitationAmount float64 `json:"precipitation_amount"`
-	} `json:"instant"`
-	Predicted struct {
-		Summary                     string  `json:"summary"`
-		Confidence                  string  `json:"confidence"`
-		AirTemperatureMax           float64 `json:"air_temperature_max"`
-		AirTemperatureMin           float64 `json:"air_temperature_min"`
-		PrecipitationAmount         float64 `json:"precipitation_amount"`
-		PrecipitationAmountMax      float64 `json:"precipitation_amount_max"`
-		PrecipitationAmountMin      float64 `json:"precipitation_amount_min"`
-		ProbabilityOfPrecipitation	float64 `json:"probability_of_precipitation"`
-	} `json:"predicted"`
-}
-
-// WeatherData structure stores current and predicted weather data for the next 9 days.
-//
-// Functionality: Handler, get
-type WeatherData struct {
-	Updated string `json:"updated"`
-	Timeseries map[string]Timeseries `json:"timeseries"`
-}
-
 // Handler will handle http request for REST service.
 func (weatherData *WeatherData) Handler(lat string, lon string) (int, error) {
-	//try to get data from database and branch if an error occurred
+	//get data from database and branch if an error occurred
 	id := lat + "&" + lon
 	data, exist := db.DB.Get("WeatherData", id)
-	withinTimeframe := false
-
-	// Check if "Time" key exists
-	if _, ok := data["Time"].(string); ok {
-		//get status on timeframe and branch if an error occurred
-		withinTimeframe, _ = db.CheckDate(data["Time"].(string), 3)
-	}
-
-	//check if data is in database and if it's usable then either read data or get new data
-	if exist && withinTimeframe {
-		err := weatherData.readData(data["Container"].(interface{}))
-		weatherData.Updated = data["Time"].(string)
+	//check if data is valid and branch if an error occurred
+	var err error
+	valid := false
+	if date, ok := data["Time"].(string); ok {
+		valid, err = db.CheckDate(date, 3)
 		if err != nil {
 			return http.StatusInternalServerError, err
 		}
+	}
+	//check if data is in database and if it's valid, then either read data or get new data
+	if exist && valid {
+		//parse data to WeatherData structure and branch if an error occurred
+		err := weatherData.readData(data["Container"].(interface{}))
+		if err != nil {
+			return http.StatusInternalServerError, err
+		}
+		weatherData.Updated = data["Time"].(string)
 	} else {
 		//get data based on coordinates and branch if an error occured
 		status, err := weatherData.get(lat, lon)
 		if err != nil {
 			return status, err
 		}
-		//send data to database
+		//send data to database and branch if an error occured 
 		var data db.Data
 		data.Container = weatherData
 		date, _, err := db.DB.Add("WeatherData", id, data)
@@ -78,20 +49,22 @@ func (weatherData *WeatherData) Handler(lat string, lon string) (int, error) {
 
 // get will get data for structure.
 func (weatherData *WeatherData) get(lat string, lon string) (int, error) {
-	var yr Yr
 	//get weather data from Yr and branch if an error occurred
+	var yr Yr
 	status, err := yr.get(lat, lon)
 	if err != nil {
 		return status, err
 	}
+	//set weather data for all available days
 	weatherData.Timeseries = make(map[string]Timeseries)
-	prevDate := ""
+	var prevDate string
 	for _, elem := range yr.Properties.Timeseries {
-		var timeEntry Timeseries
+		//only add weather data once for each day
 		rawDate, _ := time.Parse("2006-01-02T15:04:05Z", elem.Time)
 		date := rawDate.Format("2006-01-02")
 		if prevDate != date {
 			//set data in structure
+			var timeEntry Timeseries
 			timeEntry.Instant.AirTemperature = elem.Data.Instant.Details.AirTemperature
 			timeEntry.Instant.CloudAreaFraction = elem.Data.Instant.Details.CloudAreaFraction
 			timeEntry.Instant.DewPointTemperature = elem.Data.Instant.Details.DewPointTemperature
@@ -100,7 +73,6 @@ func (weatherData *WeatherData) get(lat string, lon string) (int, error) {
 			timeEntry.Instant.WindSpeed = elem.Data.Instant.Details.WindSpeed
 			timeEntry.Instant.WindSpeedOfGust = elem.Data.Instant.Details.WindSpeedOfGust
 			timeEntry.Instant.PrecipitationAmount = elem.Data.Next1Hours.Details.PrecipitationAmount
-			
 			timeEntry.Predicted.Summary = elem.Data.Next12Hours.Summary.SymbolCode
 			timeEntry.Predicted.Confidence = elem.Data.Next12Hours.Summary.SymbolConfidence
 			timeEntry.Predicted.AirTemperatureMax = elem.Data.Next6Hours.Details.AirTemperatureMax
@@ -116,14 +88,22 @@ func (weatherData *WeatherData) get(lat string, lon string) (int, error) {
 	return http.StatusOK, nil
 }
 
+// readData parses data from database to WeatherData structure format.
 func (weatherData *WeatherData) readData(data interface{}) error {
+	//parse nested structure in data and branch if an error occurred
+	var timeseries map[string]interface{}
+	rawData := data.(map[string]interface{})
+	if parsedTimeseries, ok := rawData["Timeseries"].(map[string]interface{}); ok {
+		timeseries = parsedTimeseries
+	} else {
+		return errors.New("parsing data: invalid data structure")
+	}
+	//parse weather data for all available days
 	weatherData.Timeseries = make(map[string]Timeseries)
-    rawData := data.(map[string]interface{})
-	timeseries := rawData["Timeseries"].(map[string]interface{})
-	for key, elem := range timeseries {
-		var timeEntry Timeseries
+	for date, elem := range timeseries {
 		data := elem.(map[string]interface{})
-
+		//set data in structure
+		var timeEntry Timeseries
 		instant := data["Instant"].(map[string]interface{})
 		timeEntry.Instant.AirTemperature = instant["AirTemperature"].(float64)
 		timeEntry.Instant.CloudAreaFraction = instant["CloudAreaFraction"].(float64)
@@ -133,7 +113,6 @@ func (weatherData *WeatherData) readData(data interface{}) error {
 		timeEntry.Instant.WindSpeed = instant["WindSpeed"].(float64)
 		timeEntry.Instant.WindSpeedOfGust = instant["WindSpeedOfGust"].(float64)
 		timeEntry.Instant.PrecipitationAmount = instant["PrecipitationAmount"].(float64)
-
 		predicted := data["Predicted"].(map[string]interface{})
 		timeEntry.Predicted.Summary = predicted["Summary"].(string)
 		timeEntry.Predicted.Confidence = predicted["Confidence"].(string)
@@ -143,7 +122,7 @@ func (weatherData *WeatherData) readData(data interface{}) error {
 		timeEntry.Predicted.PrecipitationAmountMax = predicted["PrecipitationAmountMax"].(float64)
 		timeEntry.Predicted.PrecipitationAmountMin = predicted["PrecipitationAmountMin"].(float64)
 		timeEntry.Predicted.ProbabilityOfPrecipitation = predicted["ProbabilityOfPrecipitation"].(float64)
-		weatherData.Timeseries[key] = timeEntry
+		weatherData.Timeseries[date] = timeEntry
 	}
 	return nil
 }